+++ CHANGES.rst
@@ -8,6 +8,7 @@ in progress
 - Fix import error on Python 3.7.
   ``AttributeError: module 'asyncio' has no attribute 'exceptions'``
 - Add more software tests
+- Add a few example programs
 
 
 2022-07-25 0.4.0